+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - [121](https://github.com/vegaprotocol/vegatools/pull/121) - Add option to `snapshotdb` to print a snapshot to a file as JSON
 - [140](https://github.com/vegaprotocol/vegatools/issue/140) - Add status column to liquidity commitment viewer
 - [156](https://github.com/vegaprotocol/vegatools/issue/156) - Upgraded go build versions to 1.16 + 1.17
+- [165](https://github.com/vegaprotocol/vegatools/issue/165) - Adding perftest subcommand
 
 ### 🐛 Fixes
 - [78](https://github.com/vegaprotocol/vegatools/pull/78) - Fix build with missing dependency